capstone-project-1-final/pyproject.toml
[project]
name = "retail-banking-rag-v1"
version = "0.1.0"
description = "Retail Banking RAG System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docling",
    "fastapi>=0.135.2",
    "langchain",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-postgres>=0.0.17",
    "psycopg2>=2.9.11",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-multipart",
    "streamlit",
    "tiktoken>=0.12.0",
    "uvicorn>=0.42.0",
]
